Shortlist of changes as follows: Reorganised folder structure.
EnemySpawning.cs substantial changes were made to this class and its mechanics, including implementation of the system to allow for in editor creation and alteration of spawned waves and spawn locations.
EnemyBase.cs minor changes including getting a reference to the player when they are spawned instead of when they are attacking or are killed by the player, this fixes a null pointer that occured when the player died as an enemy died.
Also tweaked PathB to allow for positive or negative positioning.
GameStart.cs added functionality to detect which scene is loaded to allow the timescale to be controlled.
SpawnPositions.cs Massive changes including introducing support for in editor creation and alteration of waves and spawn positions.

Added the Standard Assets Unity Pack to the project for exploring the Shop Scene later on, likely to be removed once the shop is constructed.
Player Progression Shop Progress.
Altered UV animation scripts to allow for more versitility with usage. Namely so that the Improved Missile Fuel Injectors could have an animated material.
Created a small modular pipe kit.
Created a sample trim sheet to test Unity's lighting and rendering capabilities as it has been a few years since I last used Unity's lighting engine.
Created small test scene called UpgradeShopScene.
Created Improved Missile Fuel Injector asset.